# Vector implementation for arraycmp on P8/P9