--- a/compiler/p/codegen/OMRInstOpCodeProperties.hpp
+++ b/compiler/p/codegen/OMRInstOpCodeProperties.hpp
@@ -993,6 +993,17 @@
    },
 
    {
+   /* .mnemonic    = */ OMR::InstOpCode::cmpb,
+   /* .name        = */ "cmpb",
+   /* .description =    "Compare Bytes", */
+   /* .prefix      = */ 0x00000000,
+   /* .opcode      = */ 0x7C0003F8,
+   /* .format      = */ FORMAT_RA_RS_RB,
+   /* .minimumALS  = */ OMR_PROCESSOR_PPC_P6,
+   /* .properties  = */ PPCOpProp_SyncSideEffectFree,
+   },
+
+   {
    /* .mnemonic    = */ OMR::InstOpCode::cmprb,
    /* .name        = */ "cmprb",
    /* .description =    "Compare Ranged Byte", */
@@ -8836,17 +8847,17 @@
    /*                   PPCOpProp_SyncSideEffectFree, */
    /* }, */
 
-   /* { */
-   /* .mnemonic    =    OMR::InstOpCode::vclzd, */
-   /* .name        =    "vclzd", */
+   {
+   /* .mnemonic    = */ OMR::InstOpCode::vclzd,
+   /* .name        = */ "vclzd",
    /* .description =    "Vector Count Leading Zeros Dword", */
-   /* .prefix      =    0x00000000, */
-   /* .opcode      =    0x100007C2, */
-   /* .format      =    FORMAT_UNKNOWN, */
-   /* .minimumALS  =    OMR_PROCESSOR_PPC_P8, */
-   /* .properties  =    PPCOpProp_IsVMX | */
-   /*                   PPCOpProp_SyncSideEffectFree, */
-   /* }, */
+   /* .prefix      = */ 0x00000000,
+   /* .opcode      = */ 0x100007C2,
+   /* .format      = */ FORMAT_VRT_VRB,
+   /* .minimumALS  = */ OMR_PROCESSOR_PPC_P8,
+   /* .properties  = */ PPCOpProp_IsVMX |
+                        PPCOpProp_SyncSideEffectFree,
+   },
 
    /* { */
    /* .mnemonic    =    OMR::InstOpCode::vclzh, */
@@ -8860,17 +8871,17 @@
    /*                   PPCOpProp_SyncSideEffectFree, */
    /* }, */
 
-   /* { */
-   /* .mnemonic    =    OMR::InstOpCode::vclzw, */
-   /* .name        =    "vclzw", */
+   {
+   /* .mnemonic    = */ OMR::InstOpCode::vclzw,
+   /* .name        = */ "vclzw",
    /* .description =    "Vector Count Leading Zeros Word", */
-   /* .prefix      =    0x00000000, */
-   /* .opcode      =    0x10000782, */
-   /* .format      =    FORMAT_UNKNOWN, */
-   /* .minimumALS  =    OMR_PROCESSOR_PPC_P8, */
-   /* .properties  =    PPCOpProp_IsVMX | */
-   /*                   PPCOpProp_SyncSideEffectFree, */
-   /* }, */
+   /* .prefix      = */ 0x00000000,
+   /* .opcode      = */ 0x10000782,
+   /* .format      = */ FORMAT_VRT_VRB,
+   /* .minimumALS  = */ OMR_PROCESSOR_PPC_P8,
+   /* .properties  = */ PPCOpProp_IsVMX |
+                        PPCOpProp_SyncSideEffectFree,
+   },
 
    /* { */
    /* .mnemonic    =    OMR::InstOpCode::vcmpbfp, */