WIP implementation for modifying P8/P9 arraycmp to use vector operations.