@IBMJimmyk 2023-05-29
The new implementation will use `lxvw4x` to load 16 bytes from each array at a time to perform the comparison. This is okay on both big and little endian since at this point it only needs to know if there is a difference in the 16 bytes but does not yet need to know which byte position the difference is at. If the bytes are the same, it moves on to the next 16 bytes. If there is a difference, 16 bytes are reloaded one at a time via the residue loop to identify at which position the difference occurred at. If there are fewer than 16 bytes left to load, the residue loop is also used.

The current code looks like this:
```
------------------------------
 n8763n   (  0)  treetop                                                                              [    0x74e17b95a2d0] bci=[0,10,83] rc=0 vc=4277 vn=- li=10 udi=- nc=1
 n8762n   (  3)    arraycmp  <arraycmp>[#241  helper Method] [flags 0x400 0x0 ] (in GPR_0163) (arrayCmpLen )  [    0x74e17b95a280] bci=[0,10,83] rc=3 vc=4277 vn=- li=10 udi=35760 nc=3 flg=>
 n8749n   (  0)      aladd (in GPR_0160) (X>=0 internalPtr )                                          [    0x74e17b959e70] bci=[0,10,83] rc=0 vc=4277 vn=- li=10 udi=34944 nc=2 flg=0x8100
 n33n     (  0)        ==>l2a (in &GPR_0104) (X!=0 )
 n8751n   (  0)        lconst 8 (highWordZero X!=0 X>=0 cannotOverflow )                              [    0x74e17b959f10] bci=[0,10,83] rc=0 vc=4277 vn=- li=10 udi=- nc=0 flg=0x5104
 n8755n   (  0)      aladd (in GPR_0161) (X>=0 internalPtr )                                          [    0x74e17b95a050] bci=[0,13,83] rc=0 vc=4277 vn=- li=10 udi=35216 nc=2 flg=0x8100
 n38n     (  0)        ==>l2a (in &GPR_0105)
 n8751n   (  0)        ==>lconst 8 (highWordZero X!=0 X>=0 cannotOverflow )
 n8761n   (  0)      i2l (in GPR_0162) (highWordZero X>=0 )                                           [    0x74e17b95a230] bci=[0,4,82] rc=0 vc=4277 vn=- li=10 udi=35488 nc=1 flg=0x4100
 n411n    (  2)        ==>iloadi (in GPR_0112) (X>=0 cannotOverflow )
------------------------------

 [    0x74e0522588f0]   10      addi    GPR_0160, &GPR_0104, 8
 [    0x74e052258a00]   13      addi    GPR_0161, &GPR_0105, 8
 [    0x74e052258b10]   4       extsw   GPR_0162, GPR_0112
 [    0x74e052259540]   10      Label L0337:    ; (Start of internal control flow)
 [    0x74e0522595d0]   10      li      GPR_0163, 0000000000000000
 [    0x74e052259670]   10      cmpdi   CCR_0168, GPR_0162, 16
 [    0x74e052259710]   10      blt     CCR_0168, Label L0339
 [    0x74e0522597b0]   10      sradi   GPR_0164, GPR_0162, 4
 [    0x74e052259850]   10      mtctr   GPR_0164
 [    0x74e0522598f0]   10      Label L0338:
 [    0x74e052259980]   10      lxvw4x  VRF_0166, GPR_0163, GPR_0160
 [    0x74e052259a20]   10      lxvw4x  VRF_0167, GPR_0163, GPR_0161
 [    0x74e052259ac0]   10      vcmpequb.       VRF_0166, VRF_0166, VRF_0167
 [    0x74e052259b60]   10      bge     CCR_0168, Label L0340
 [    0x74e052259c00]   10      addi    GPR_0163, GPR_0163, 16
 [    0x74e052259ca0]   10      bdnz    CCR_0168, Label L0338
 [    0x74e052259d40]   10      Label L0339:
 [    0x74e052259dd0]   10      cmpd    CCR_0168, GPR_0163, GPR_0162
 [    0x74e052259e70]   10      beq     CCR_0168, Label L0341
 [    0x74e052259f10]   10      Label L0340:
 [    0x74e052259fa0]   10      lbzx    GPR_0164, GPR_0163, GPR_0160
 [    0x74e05225a040]   10      lbzx    GPR_0165, GPR_0163, GPR_0161
 [    0x74e05225a0e0]   10      cmpw    CCR_0168, GPR_0164, GPR_0165
 [    0x74e05225a180]   10      bne     CCR_0168, Label L0342
 [    0x74e05225a220]   10      addi    GPR_0163, GPR_0163, 1
 [    0x74e05225a2c0]   10      cmpd    CCR_0168, GPR_0163, GPR_0162
 [    0x74e05225a360]   10      bne     CCR_0168, Label L0340
 [    0x74e05225a400]   10      Label L0341:
 [    0x74e05225a490]   10      Label L0342:
 [    0x74e05225a520]   10      Label L0343:    ; (End of internal control flow)
```

I wrote a small test to perform `arraycmp` 100 million times on various array sizes.

| Size | Baseline (s) | Vector (s) | Diff |
| --- | --- | --- | --- |
| 1 | 0.482 | 0.414 | 0.859 |
| 2 | 0.576 | 0.498 | 0.865 |
| 3 | 0.755 | 0.816 | 1.081 |
| 4 | 0.763 | 0.698 | 0.915 |
| 5 | 0.872 | 0.832 | 0.954 |
| 6 | 0.999 | 0.975 | 0.976 |
| 7 | 1.111 | 1.089 | 0.980 |
| 8 | 0.617 | 1.238 | 2.006 |
| 9 | 0.825 | 1.375 | 1.667 |
| 10 | 0.953 | 1.52 | 1.595 |
| 11 | 1.044 | 2.077 | 1.989 |
| 12 | 1.153 | 2.207 | 1.914 |
| 13 | 1.282 | 2.436 | 1.900 |
| 14 | 1.439 | 2.56 | 1.779 |
| 15 | 1.566 | 2.576 | 1.645 |
| 16 | 0.763 | 0.441 | 0.578 |
| 17 | 0.959 | 0.524 | 0.546 |
| 18 | 1.051 | 0.61 | 0.580 |
| 19 | 1.136 | 0.992 | 0.873 |
| 20 | 1.245 | 0.819 | 0.658 |
| 21 | 1.406 | 0.976 | 0.694 |
| 22 | 1.531 | 1.011 | 0.660 |
| 23 | 1.667 | 1.11 | 0.666 |
| 24 | 0.83 | 1.232 | 1.484 |
| 25 | 1.073 | 1.509 | 1.406 |
| 26 | 1.203 | 1.767 | 1.469 |
| 27 | 1.248 | 1.888 | 1.513 |
| 28 | 1.357 | 2.428 | 1.789 |
| 29 | 1.634 | 2.676 | 1.638 |
| 30 | 1.648 | 2.8 | 1.699 |
| 31 | 1.786 | 2.798 | 1.567 |
| 32 | 0.968 | 0.892 | 0.921 |

@zl-wang You mentioned you wanted to take a look at this.