@zl-wang 2023-05-29
that is too much performance difference.  i think a few improvement should be pursued:
1) a separate implementation for POWER9 which supports both lxvh8x/vcmpequh/vclzh and lxvb16x/vcmpequb/vclzb (although missing vector load with length ... i.e. still need a residue loop for the case of comparing equal in the vector portion)
2) POWER8 also has vclz(h|b) to make it convenient for calculating vector comparison portion;
3) residue loop can tailor to char or byte cases.  not always using byte load, that is;