@zl-wang 2023-05-29
for typical String comparison, if using char (2-byte) and residue iteration is going to be half of the current, i expected no case is going to be 2x slower.