@zl-wang 2023-05-29
actually, i projected most of the cases would be faster.  only 8-byte case could be slower, but almost equal.