@IBMJimmyk 2023-06-12
The newest change adds an earlier length check against 4 after the vector load section. This will skip past both the 8 byte and 4 byte load sections instead of just skipping over one at a time.

The latest implementation looks like this:
```
startlabel:
  li       offsetReg, 0
  cmpdi    cr6, lengthReg, 4
  blt      cr6, skipLoad4Label
  cmpdi    cr6, lengthReg, 8
  blt      cr6, skipLoad8Label
  cmpdi    cr6, lengthReg, 16
  blt      cr6, skipLoad16Label
  sradi    tempReg, lengthReg, 4
  mtctr    tempReg
load16LoopLabel:
  lxvw4x   vec1Reg, offsetReg, src1AddrReg
  lxvw4x   vec2Reg, offsetReg, src2AddrReg
  vcmpequb vec1Reg, vec1Reg, vec2Reg
  bge      cr6, load16DiffLabel
  addi     offsetReg, offsetReg, 16
  bdnz     load16LoopLabel
  b        doneLoad16Label
load16DiffLabel:
#if (ppcle)
  vspltisw vec2Reg, -16
  vrlw     vec1Reg, vec1Reg, vec2Reg
  vspltish vec2Reg, 8
  vrlh     vec1Reg, vec1Reg, vec2Reg
#endif
  mfvsrd   tempReg, vec1Reg
  nor      tempReg, tempReg, tempReg
  cntlzd   tempReg, tempReg
  sradi    tempReg, tempReg, 3
  add      offsetReg, offsetReg, tempReg
  cmpdi    cr6, tempReg, 8
  bne      cr6, returnVectorFalseLabel
  xxpermdi vec1Reg, vec1Reg, vec1Reg, 3  /*  use mfvsrld for Power 9
  mfvsrd   tempReg, vec1Reg               */
  nor      tempReg, tempReg, tempReg
  cntlzd   tempReg, tempReg
  sradi    tempReg, tempReg, 3
  add      offsetReg, offsetReg, tempReg
  b        returnVectorFalseLabel
load4DiffLabel:
  addi     offsetReg, offsetReg, -4
load8DiffLabel:
  cmpb     tempReg, tempReg, tempReg2
  nor      tempReg, tempReg, tempReg
  cntlzd   tempReg, tempReg
  sradi    tempReg, tempReg, 3
  add      offsetReg, offsetReg, tempReg
  b        returnVectorFalseLabel
doneLoad16Label:
  subf.    tempReg, offsetReg, lengthReg
  cmpdi    cr6, tempReg, 4
  blt      cr6, skipLoad4Label2
  cmpdi    cr6, tempReg, 8
  blt      cr6, skipLoad8Label2
skipLoad16Label:
  ldbrx    tempReg, offsetReg, src1AddrReg    //ldx for big endian
  ldbrx    tempReg2, offsetReg, src2AddrReg   //ldx for big endian
  cmpd     cr6, tempReg, tempReg2
  bne      cr6, load8DiffLabel
  addi     offsetReg, offsetReg, 8
  subf.    tempReg, offsetReg, lengthReg
skipLoad8Label2:
  cmpdi    cr6, tempReg, 4
  blt      cr6, skipLoad4Label2
skipLoad8Label:
  lwbrx    tempReg, offsetReg, src1AddrReg    //lwx for big endian
  lwbrx    tempReg2, offsetReg, src2AddrReg   //lwx for big endian
  cmpw     cr6, tempReg, tempReg2
  bne      cr6, load4DiffLabel
  addi     offsetReg, offsetReg, 4
skipLoad4Label:
  subf.    tempReg, offsetReg, lengthReg
skipLoad4Label2:
  beq      cr0, returnTrueLabel
  mtctr    tempReg
residueLoopLabel:
  lbzx     tempReg, offsetReg, src1AddrReg
  lbzx     tempReg2, offsetReg, src2AddrReg
  cmpw     cr6, tempReg, tempReg2
  bne      cr6, returnFalseLabel
  addi     offsetReg, offsetReg, 1
  bdnz     residueLoopLabel
returnTrueLabel:
returnVectorFalseLabel:
returnFalseLabel:
```
There is a vector 16 byte load section followed by an 8 byte load, 4 byte load then the residue loop. If a mismatch is found in the vector section, it will jump to `load16DiffLabel` to locate where in the vector register the difference occured. If it happens in the 8 byte or 4 byte load sections, `cmpb` is used to locate the difference.

I also updated the performance data. The number of interations to 1 billion which is the same as the previous tests.
These runs were done on the perf machine `cit926`.
I added more test cases so I had to retake the baseline.

This table covers the cases where there is a complete match on arrays of different lengths:
| Length | Baseline (s) | Vector (s) | Diff% |
| --- | --- | --- | --- |
| 1 | 3.63 | 3.092 | 17.4% |
| 2 | 4.366 | 3.783 | 15.4% |
| 3 | 5.749 | 5.85 | -1.7% |
| 4 | 5.815 | 3.24 | 79.5% |
| 5 | 6.64 | 3.835 | 73.1% |
| 6 | 7.725 | 4.374 | 76.6% |
| 7 | 8.441 | 5.079 | 66.2% |
| 8 | 4.818 | 3.879 | 24.2% |
| 9 | 6.442 | 4.505 | 43% |
| 10 | 7.281 | 6.936 | 5% |
| 11 | 7.986 | 6.004 | 33% |
| 12 | 8.806 | 4.494 | 96% |
| 13 | 9.832 | 4.938 | 99.1% |
| 14 | 11.046 | 7.333 | 50.6% |
| 15 | 12.158 | 7.146 | 70.1% |
| 16 | 5.619 | 5.331 | 5.4% |
| 17 | 7.335 | 5.763 | 27.3% |
| 18 | 8.231 | 8.484 | -3% |
| 19 | 8.819 | 7.057 | 25% |
| 20 | 9.652 | 6.163 | 56.6% |
| 21 | 11.239 | 6.578 | 70.9% |
| 22 | 11.831 | 7.154 | 65.4% |
| 23 | 12.893 | 7.791 | 65.5% |
| 24 | 6.328 | 6.174 | 2.5% |
| 25 | 8.104 | 6.588 | 23% |
| 26 | 8.764 | 7.067 | 24% |
| 27 | 9.592 | 7.788 | 23.2% |
| 28 | 10.417 | 6.834 | 52.4% |
| 29 | 12.392 | 7.103 | 74.5% |
| 30 | 12.51 | 7.712 | 62.2% |
| 31 | 13.752 | 8.205 | 67.6% |
| 32 | 7.3 | 7.601 | -4% |
| 39 | 14.338 | 9.408 | 52.4% |
| 40 | 8.4 | 8.143 | 3.2% |
| 47 | 15.142 | 9.624 | 57.3% |
| 48 | 9.13 | 8.593 | 6.3% |
| 55 | 16.064 | 10.823 | 48.4% |
| 56 | 9.971 | 8.793 | 13.4% |
| 63 | 16.738 | 11.065 | 51.3% |
| 64 | 10.572 | 9.557 | 10.6% |
| 71 | 17.731 | 11.724 | 51.2% |
| 72 | 11.375 | 10.055 | 13.1% |
| 79 | 18.525 | 12.027 | 54% |
| 80 | 12.348 | 10.927 | 13% |
| 87 | 19.244 | 13.025 | 47.8% |
| 88 | 13.072 | 11.923 | 9.6% |
| 95 | 20.238 | 13.07 | 54.8% |
| 96 | 17.385 | 11.561 | 50.4% |
| 103 | 23.483 | 13.707 | 71.3% |
| 104 | 19.293 | 12.12 | 59.2% |
| 111 | 24.383 | 14.136 | 72.5% |
| 112 | 20.071 | 13.21 | 51.9% |
| 119 | 25.211 | 15.599 | 61.6% |
| 120 | 19.999 | 13.516 | 48% |
| 127 | 26.104 | 15.686 | 66.4% |
| 128 | 20.762 | 13.873 | 49.7% |
| 1000 | 114.716 | 82.653 | 38.8% |

The vector implementation helps in most case. The worst case seems to be 4% down on a 32 byte array. That path uses two vector loads to handle the entire match so I'm not entirely sure why it seems to be worse.

The following data is for the case where there a mismatched byte in the two arrays being compared.
There is a length parameter that is the length of the array and also this index of the mismatched element. This goes from 0 to length-1. So for a 7 length array it goes from 0-6.

Different array lengths take different paths through the code.
3 length array - goes stright to the byte by byte residue loop.
7 length array - uses a 4 byte load for the first 4 bytes then goes to the byte by byte residue loop.
11 length array - uses an 8 byte load for the first 8 byte then goes to the byte by byte residue loop.
15 length array - uses and 8 byte load followed by an 4 byte load then the residue loop
19 length array - uses a 16 byte vector load for the first 16 bytes then goes to the byte by byte residue loop.
23 length array - uses 16 byte load then 4 byte load
27 length array - uses 16 byte load then 8 byte load
31 length array - uses 16 byte load then 8 byte load then 4 byte load

| Length | Mismatch Idx | Baseline (s) | Vector (s) | Diff% |
| --- | --- | --- | --- | --- |
| 3 | 0 | 4.604 | 4.492 | 2.5% |
| 3 | 1 | 6.14 | 5.326 | 15.3% |
| 3 | 2 | 6.112 | 6.063 | 0.8% |
| 7 | 0 | 4.591 | 5.263 | -12.8% |
| 7 | 1 | 6.303 | 5.262 | 19.8% |
| 7 | 2 | 6.109 | 5.356 | 14.1% |
| 7 | 3 | 6.951 | 5.267 | 32% |
| 7 | 4 | 7.859 | 4.48 | 75.4% |
| 7 | 5 | 8.775 | 5.35 | 64% |
| 7 | 6 | 9.851 | 6.211 | 58.6% |
| 11 | 0 | 6.285 | 5.56 | 13% |
| 11 | 1 | 7.248 | 5.567 | 30.2% |
| 11 | 2 | 7.964 | 5.465 | 45.7% |
| 11 | 3 | 8.922 | 5.472 | 63.1% |
| 11 | 4 | 9.849 | 5.565 | 77% |
| 11 | 5 | 10.992 | 5.565 | 97.5% |
| 11 | 6 | 11.99 | 5.47 | 119.2% |
| 11 | 7 | 13.029 | 5.57 | 133.9% |
| 11 | 8 | 6.324 | 5.498 | 15% |
| 11 | 9 | 7.397 | 6.376 | 16% |
| 11 | 10 | 7.924 | 7.11 | 11.5% |
| 15 | 0 | 6.164 | 5.468 | 12.7% |
| 15 | 1 | 7.102 | 5.561 | 27.7% |
| 15 | 2 | 8.079 | 5.566 | 45.2% |
| 15 | 3 | 8.905 | 5.469 | 62.8% |
| 15 | 4 | 9.974 | 5.571 | 79% |
| 15 | 5 | 10.916 | 5.47 | 99.6% |
| 15 | 6 | 11.987 | 5.468 | 119.2% |
| 15 | 7 | 12.911 | 5.466 | 136.2% |
| 15 | 8 | 6.459 | 5.761 | 12.1% |
| 15 | 9 | 7.255 | 5.762 | 25.9% |
| 15 | 10 | 8.049 | 5.762 | 39.7% |
| 15 | 11 | 8.815 | 5.762 | 53% |
| 15 | 12 | 10.926 | 5.353 | 104.1% |
| 15 | 13 | 10.96 | 5.997 | 82.8% |
| 15 | 14 | 12.133 | 6.993 | 73.5% |
| 19 | 0 | 6.146 | 6.476 | -5.1% |
| 19 | 1 | 7.108 | 6.517 | 9.1% |
| 19 | 2 | 8.081 | 6.461 | 25.1% |
| 19 | 3 | 8.902 | 6.585 | 35.2% |
| 19 | 4 | 9.975 | 6.587 | 51.4% |
| 19 | 5 | 10.992 | 6.585 | 66.9% |
| 19 | 6 | 11.987 | 6.586 | 82% |
| 19 | 7 | 12.918 | 6.589 | 96.1% |
| 19 | 8 | 7.762 | 7.291 | 6.5% |
| 19 | 9 | 7.95 | 7.286 | 9.1% |
| 19 | 10 | 8.842 | 7.158 | 23.5% |
| 19 | 11 | 9.879 | 7.297 | 35.4% |
| 19 | 12 | 10.8 | 7.159 | 50.9% |
| 19 | 13 | 11.839 | 7.286 | 62.5% |
| 19 | 14 | 12.864 | 7.283 | 76.6% |
| 19 | 15 | 14.26 | 7.157 | 99.3% |
| 19 | 16 | 7.297 | 6.589 | 10.8% |
| 19 | 17 | 7.975 | 7.126 | 11.9% |
| 19 | 18 | 8.665 | 8.019 | 8.1% |
| 23 | 0 | 6.286 | 6.478 | -3% |
| 23 | 1 | 7.117 | 6.477 | 9.9% |
| 23 | 2 | 7.943 | 6.515 | 21.9% |
| 23 | 3 | 9.022 | 6.586 | 37% |
| 23 | 4 | 9.873 | 6.588 | 49.9% |
| 23 | 5 | 10.913 | 6.586 | 65.7% |
| 23 | 6 | 11.962 | 6.612 | 80.9% |
| 23 | 7 | 13.029 | 6.608 | 97.2% |
| 23 | 8 | 7.902 | 7.284 | 8.5% |
| 23 | 9 | 8.084 | 7.3 | 10.7% |
| 23 | 10 | 8.846 | 7.281 | 21.5% |
| 23 | 11 | 9.694 | 7.156 | 35.5% |
| 23 | 12 | 10.674 | 7.175 | 48.8% |
| 23 | 13 | 11.721 | 7.16 | 63.7% |
| 23 | 14 | 12.727 | 7.158 | 77.8% |
| 23 | 15 | 14.258 | 7.159 | 99.2% |
| 23 | 16 | 7.292 | 7.599 | -4% |
| 23 | 17 | 8.128 | 7.604 | 6.9% |
| 23 | 18 | 8.673 | 7.47 | 16.1% |
| 23 | 19 | 9.644 | 7.473 | 29.1% |
| 23 | 20 | 10.715 | 7.134 | 50.2% |
| 23 | 21 | 11.812 | 7.786 | 51.7% |
| 23 | 22 | 13.697 | 8.613 | 59% |
| 27 | 0 | 6.188 | 6.11 | 1.3% |
| 27 | 1 | 7.129 | 6.109 | 16.7% |
| 27 | 2 | 7.958 | 6.176 | 28.9% |
| 27 | 3 | 8.949 | 6.226 | 43.7% |
| 27 | 4 | 9.896 | 6.111 | 61.9% |
| 27 | 5 | 10.946 | 6.1 | 79.4% |
| 27 | 6 | 12.127 | 6.246 | 94.2% |
| 27 | 7 | 13.051 | 6.231 | 109.5% |
| 27 | 8 | 7.834 | 6.962 | 12.5% |
| 27 | 9 | 7.97 | 6.977 | 14.2% |
| 27 | 10 | 8.842 | 6.971 | 26.8% |
| 27 | 11 | 9.717 | 6.967 | 39.5% |
| 27 | 12 | 10.816 | 6.964 | 55.3% |
| 27 | 13 | 11.752 | 6.806 | 72.7% |
| 27 | 14 | 12.882 | 6.968 | 84.9% |
| 27 | 15 | 14.16 | 6.975 | 103% |
| 27 | 16 | 8.149 | 6.83 | 19.3% |
| 27 | 17 | 9.743 | 6.699 | 45.4% |
| 27 | 18 | 9.682 | 6.834 | 41.7% |
| 27 | 19 | 10.714 | 6.843 | 56.6% |
| 27 | 20 | 11.656 | 6.69 | 74.2% |
| 27 | 21 | 12.666 | 6.693 | 89.2% |
| 27 | 22 | 14.128 | 6.712 | 110.5% |
| 27 | 23 | 15.739 | 6.747 | 133.3% |
| 27 | 24 | 8.021 | 6.828 | 17.5% |
| 27 | 25 | 8.714 | 7.761 | 12.3% |
| 27 | 26 | 9.664 | 8.495 | 13.8% |
| 31 | 0 | 6.31 | 6.187 | 2% |
| 31 | 1 | 7.145 | 6.232 | 14.7% |
| 31 | 2 | 7.972 | 6.233 | 27.9% |
| 31 | 3 | 9.053 | 6.197 | 46.1% |
| 31 | 4 | 9.875 | 6.095 | 62% |
| 31 | 5 | 11.029 | 6.237 | 76.8% |
| 31 | 6 | 12.008 | 6.112 | 96.5% |
| 31 | 7 | 13.05 | 6.188 | 110.9% |
| 31 | 8 | 7.834 | 6.809 | 15.1% |
| 31 | 9 | 8.099 | 6.962 | 16.3% |
| 31 | 10 | 8.863 | 6.808 | 30.2% |
| 31 | 11 | 9.884 | 6.811 | 45.1% |
| 31 | 12 | 10.819 | 6.809 | 58.9% |
| 31 | 13 | 11.72 | 6.81 | 72.1% |
| 31 | 14 | 12.883 | 6.964 | 85% |
| 31 | 15 | 14.167 | 6.964 | 103.4% |
| 31 | 16 | 7.996 | 6.693 | 19.5% |
| 31 | 17 | 9.832 | 6.709 | 46.6% |
| 31 | 18 | 9.875 | 6.836 | 44.5% |
| 31 | 19 | 10.714 | 6.833 | 56.8% |
| 31 | 20 | 11.559 | 6.829 | 69.3% |
| 31 | 21 | 12.667 | 6.835 | 85.3% |
| 31 | 22 | 14.126 | 6.831 | 106.8% |
| 31 | 23 | 15.632 | 6.71 | 133% |
| 31 | 24 | 8.022 | 7.586 | 5.8% |
| 31 | 25 | 8.862 | 7.411 | 19.6% |
| 31 | 26 | 9.656 | 7.408 | 30.4% |
| 31 | 27 | 10.475 | 7.574 | 38.3% |
| 31 | 28 | 11.621 | 7.255 | 60.2% |
| 31 | 29 | 12.8 | 7.872 | 62.6% |
| 31 | 30 | 14.03 | 8.483 | 65.4% |

The big problem here is being 13% slower when the array is 7 bytes long and the mismatch is on the very first byte. Under the old implementation, it would go straight to the residue loop and get a hit on the first compare. Under the new implementation it attempts to use a 4 byte load, finds a mismatch then takes a path to find the mismatch within the 4 bytes. The next biggest problem is being 5% down on a 19 byte array when there is a mismatch on the first byte. This one needs to do the vector mismatch path to find a mismatch on the first byte.