@IBMJimmyk 2023-06-19
With my latest changes the assembly now looks like this:
```
startlabel:
  li       offsetReg, 0
  cmpdi    cr6, lengthReg, 4
  blt      cr6, skipLoad4Label
  cmpdi    cr6, lengthReg, 8
  blt      cr6, skipLoad8Label
  cmpdi    cr6, lengthReg, 16
  blt      cr6, skipLoad16Label
  sradi    tempReg, lengthReg, 4
  mtctr    tempReg
load16LoopLabel:
  lxvw4x   vec1Reg, offsetReg, src1AddrReg  // use lxvb16x instruction on P9
  lxvw4x   vec2Reg, offsetReg, src2AddrReg  // use lxvb16x instruction on P9
  vcmpequb vec1Reg, vec1Reg, vec2Reg
  bge      cr6, load16DiffLabel
  addi     offsetReg, offsetReg, 16
  bdnz     load16LoopLabel
  subf.    tempReg, offsetReg, lengthReg
  beq      cr0, returnTrueLabel
  cmpdi    cr6, tempReg, 4
  blt      cr6, skipLoad4Label2
  cmpdi    cr6, tempReg, 8
  blt      cr6, skipLoad8Label
  b        skipLoad16Label
load16DiffLabel:
#if (ppcle)
  vspltisw vec2Reg, -16
  vrlw     vec1Reg, vec1Reg, vec2Reg
  vspltish vec2Reg, 8
  vrlh     vec1Reg, vec1Reg, vec2Reg
#endif
  vnor     vec1Reg, vec1Reg, vec1Reg
  vclzd    vec1Reg, vec1Reg
  mfvsrd   tempReg, vec1Reg
  sradi    tempReg, tempReg, 3
  add      offsetReg, offsetReg, tempReg
  cmpdi    cr6, tempReg, 8
  bne      cr6, returnVectorFalseLabel
  xxpermdi vec1Reg, vec1Reg, vec1Reg, 3  /*  use mfvsrld for Power 9
  mfvsrd   tempReg, vec1Reg               */
  sradi    tempReg, tempReg, 3
  add      offsetReg, offsetReg, tempReg
  b        returnVectorFalseLabel
load4DiffLabel:
  addi     offsetReg, offsetReg, -4
load8DiffLabel:
  cmpb     tempReg, tempReg, tempReg2
  nor      tempReg, tempReg, tempReg
  cntlzd   tempReg, tempReg
  sradi    tempReg, tempReg, 3
  add      offsetReg, offsetReg, tempReg
  b        returnVectorFalseLabel
skipLoad16Label:
  ldbrx    tempReg, offsetReg, src1AddrReg    //ldx for big endian
  ldbrx    tempReg2, offsetReg, src2AddrReg   //ldx for big endian
  cmpd     cr6, tempReg, tempReg2
  bne      cr6, load8DiffLabel
  addi     offsetReg, offsetReg, 8
  subf.    tempReg, offsetReg, lengthReg
  beq      cr0, returnTrueLabel
  cmpdi    cr6, tempReg, 4
  blt      cr6, skipLoad4Label2
skipLoad8Label:
  lwbrx    tempReg, offsetReg, src1AddrReg    //lwx for big endian
  lwbrx    tempReg2, offsetReg, src2AddrReg   //lwx for big endian
  cmpw     cr6, tempReg, tempReg2
  bne      cr6, load4DiffLabel
  addi     offsetReg, offsetReg, 4
skipLoad4Label:
  subf.    tempReg, offsetReg, lengthReg
  beq      cr0, returnTrueLabel
skipLoad4Label2:
  lbzx     tempReg, offsetReg, src1AddrReg
  lbzx     tempReg2, offsetReg, src2AddrReg
  cmpw     cr6, tempReg, tempReg2
  bne      cr6, returnFalseLabel
  addi     offsetReg, offsetReg, 1
  cmpd     cr6, offsetReg, lengthReg
  beq      cr6, returnTrueLabel
  lbzx     tempReg, offsetReg, src1AddrReg
  lbzx     tempReg2, offsetReg, src2AddrReg
  cmpw     cr6, tempReg, tempReg2
  bne      cr6, returnFalseLabel
  addi     offsetReg, offsetReg, 1
  cmpd     cr6, offsetReg, lengthReg
  beq      cr6, returnTrueLabel
  lbzx     tempReg, offsetReg, src1AddrReg
  lbzx     tempReg2, offsetReg, src2AddrReg
  cmpw     cr6, tempReg, tempReg2
  bne      cr6, returnFalseLabel
  addi     offsetReg, offsetReg, 1
returnTrueLabel:
returnVectorFalseLabel:
returnFalseLabel:
```
The key change is this section just after the vector load loop:
```
  bdnz     load16LoopLabel
  subf.    tempReg, offsetReg, lengthReg
  beq      cr0, returnTrueLabel
  cmpdi    cr6, tempReg, 4
  blt      cr6, skipLoad4Label2
  cmpdi    cr6, tempReg, 8
  blt      cr6, skipLoad8Label
  b        skipLoad16Label
```
It used to just do an unconditional branch after the `bdnz` then do the checks to see which load size to use next and then branch to that section. In many cases, this just adds an extra jump. Now, it just performs the legnth checks earlier and only had to do 1 jump instead of 2.

This change also gets rid of the padding `nop` instructions since they didn't help anymore.

This changed helps in most of the places I expected it to, but I also saw more unusual results.

| Length | Mismatch Idx | Previous Vector (s) | New Vector (s) | Diff% |
| --- | --- | --- | --- | --- |
| 23 | 4 | 5.993 | 6.176 | -3.0% |
| 27 | 4 | 5.693 | 6.465 | -11.9% |

A 23 length array with a mismatch at index 4 will first use a vector load to load the first 16 byte. It will see one of the bytes a different and go down the path to figure out which byte is different.

A 27 length array with a mismatch at index 4 will do exactly the same thing. I verified this in gdb by steping through the generate assembly for the arraycmp. The array being 4 bytes longer shouldn't change anything.

Despite this, there is a clear difference in performance when going from a 23 byte array to a 27 byte array and it doesn't make much sense to me.

These are two of what seem to be several cases of results not really making sense. So while this new change seems like it should help, it is hard to tell if it actaully is.